+++ tests/test_cases/test_admin_api/test_admin_salary_api.py
@@ -1,8 +1,6 @@
-# -------------------------------------------------Test User API ---------------------------------------------------
+# -----------------------------------Test User API -----------------------------------
 def test_user_get_all_my_salary_success(client, admin_user, read_json):
-    response = client.get(
-        "/user/my/salary/", headers={"Authorization": f"Bearer {admin_user}"}
-    )
+    response = client.get("/user/my/salary/", headers={"Authorization": f"Bearer {admin_user}"})
     expected = read_json("expected_responses/admin/salary/get_all_my_salary_admin.json")
     assert response.status_code == 200
     assert response.json() == expected
@@ -12,9 +10,7 @@ def test_user_get_my_salary_by_year_success(client, admin_user, read_json):
     response = client.get(
         "/user/my/salary/year/2025", headers={"Authorization": f"Bearer {admin_user}"}
     )
-    expected = read_json(
-        "expected_responses/admin/salary/get_my_salary_by_year_admin.json"
-    )
+    expected = read_json("expected_responses/admin/salary/get_my_salary_by_year_admin.json")
     assert response.status_code == 200
     assert response.json() == expected
 
@@ -24,28 +20,20 @@ def test_user_get_my_salary_by_year_not_found(client, admin_user, read_json):
         "/user/my/salary/year/2028", headers={"Authorization": f"Bearer {admin_user}"}
     )
     assert response.status_code == 404
-    assert response.json() == {
-        "detail": "Salary record not found for employee in year 2028"
-    }
+    assert response.json() == {"detail": "Salary record not found for employee in year 2028"}
 
 
-# -------------------------------------------------Test Admin API ---------------------------------------------------
-def test_admin_admin_access_get_employee_salary_by_years_success(
-    client, admin_user, read_json
-):
+# ------------------------------Test Admin API -------------------------------
+def test_admin_admin_access_get_employee_salary_by_years_success(client, admin_user, read_json):
     response = client.get(
         "admin/salaries/year/2025", headers={"Authorization": f"Bearer {admin_user}"}
     )
-    expected = read_json(
-        "expected_responses/admin/salary/get_employees_salary_by_year_admin.json"
-    )
+    expected = read_json("expected_responses/admin/salary/get_employees_salary_by_year_admin.json")
     assert response.status_code == 200
     assert response.json() == expected
 
 
-def test_admin_admin_access_get_employee_salary_by_years_not_found(
-    client, admin_user, read_json
-):
+def test_admin_admin_access_get_employee_salary_by_years_not_found(client, admin_user, read_json):
     response = client.get(
         "admin/salaries/year/2028", headers={"Authorization": f"Bearer {admin_user}"}
     )
@@ -53,22 +41,16 @@ def test_admin_admin_access_get_employee_salary_by_years_not_found(
     assert response.json() == {"detail": "No salary records found for year 2028"}
 
 
-def test_admin_admin_access_get_all_employees_salary_success(
-    client, admin_user, read_json
-):
+def test_admin_admin_access_get_all_employees_salary_success(client, admin_user, read_json):
     response = client.get(
         "admin/salaries/year/2025", headers={"Authorization": f"Bearer {admin_user}"}
     )
-    expected = read_json(
-        "expected_responses/admin/salary/get_all_employees_salary_admin.json"
-    )
+    expected = read_json("expected_responses/admin/salary/get_all_employees_salary_admin.json")
     assert response.status_code == 200
     assert response.json() == expected
 
 
-def test_admin_admin_access_delete_employee_salary_by_empid_and_year_success(
-    client, admin_user
-):
+def test_admin_admin_access_delete_employee_salary_by_empid_and_year_success(client, admin_user):
     response = client.delete(
         "admin/salaries/employee/1/year/2025",
         headers={"Authorization": f"Bearer {admin_user}"},
@@ -76,22 +58,16 @@ def test_admin_admin_access_delete_employee_salary_by_empid_and_year_success(
     assert response.status_code == 204
 
 
-def test_admin_admin_access_delete_employee_salary_by_empid_and_year_not_found(
-    client, admin_user
-):
+def test_admin_admin_access_delete_employee_salary_by_empid_and_year_not_found(client, admin_user):
     response = client.delete(
         "admin/salaries/employee/1/year/2028",
         headers={"Authorization": f"Bearer {admin_user}"},
     )
     assert response.status_code == 404
-    assert response.json() == {
-        "detail": "Salary record not found for employee in year 2028"
-    }
+    assert response.json() == {"detail": "Salary record not found for employee in year 2028"}
 
 
-def test_admin_admin_access_get_salary_by_empid_and_year_success(
-    client, admin_user, read_json
-):
+def test_admin_admin_access_get_salary_by_empid_and_year_success(client, admin_user, read_json):
     response = client.get(
         "admin/salaries/employee/5/year/2025",
         headers={"Authorization": f"Bearer {admin_user}"},
@@ -103,17 +79,13 @@ def test_admin_admin_access_get_salary_by_empid_and_year_success(
     assert response.json() == expected
 
 
-def test_admin_admin_access_get_salary_by_empid_and_year_not_exist(
-    client, admin_user, read_json
-):
+def test_admin_admin_access_get_salary_by_empid_and_year_not_exist(client, admin_user, read_json):
     response = client.get(
         "admin/salaries/employee/100/year/2026",
         headers={"Authorization": f"Bearer {admin_user}"},
     )
     assert response.status_code == 404
-    assert response.json() == {
-        "detail": "Salary record not found for employee in year 2026"
-    }
+    assert response.json() == {"detail": "Salary record not found for employee in year 2026"}
 
 
 def test_admin_admin_access_get_salary_by_empid_success(client, admin_user, read_json):
@@ -121,16 +93,12 @@ def test_admin_admin_access_get_salary_by_empid_success(client, admin_user, read
         "admin/salaries/employee/5/salaries",
         headers={"Authorization": f"Bearer {admin_user}"},
     )
-    expected = read_json(
-        "expected_responses/admin/salary/get_emoloyee_salary_by_empid_admin.json"
-    )
+    expected = read_json("expected_responses/admin/salary/get_emoloyee_salary_by_empid_admin.json")
     assert response.status_code == 200
     assert response.json() == expected
 
 
-def test_admin_admin_access_get_salary_by_empid_not_exist(
-    client, admin_user, read_json
-):
+def test_admin_admin_access_get_salary_by_empid_not_exist(client, admin_user, read_json):
     response = client.get(
         "admin/salaries/employee/100/salaries",
         headers={"Authorization": f"Bearer {admin_user}"},
@@ -158,22 +126,16 @@ def test_admin_admin_access_create_salary_conflict(client, admin_user, read_json
     )
     assert response.status_code == 409
     assert response.json() == {
-        "detail": "Salary record already exists for employee 1 in year 2025"
+        "detail": ["Salary record already exists for employee 1 in year 2025"]
     }
 
 
-def test_admin_admin_access_delete_salary_by_salaryid_success(
-    client, admin_user, read_json
-):
-    response = client.delete(
-        "admin/salaries/1", headers={"Authorization": f"Bearer {admin_user}"}
-    )
+def test_admin_admin_access_delete_salary_by_salaryid_success(client, admin_user, read_json):
+    response = client.delete("admin/salaries/1", headers={"Authorization": f"Bearer {admin_user}"})
     assert response.status_code == 204
 
 
-def test_admin_admin_access_delete_salary_by_salaryid_not_found(
-    client, admin_user, read_json
-):
+def test_admin_admin_access_delete_salary_by_salaryid_not_found(client, admin_user, read_json):
     response = client.delete(
         "admin/salaries/100", headers={"Authorization": f"Bearer {admin_user}"}
     )
