+++ tests/test_cases/test_manager_api/test_manager_leave_api.py
@@ -1,11 +1,7 @@
-# -------------------------------------------------Test User API ---------------------------------------------------
+# -----------------------------------Test User API -----------------------------------
 def test_manager_get_my_all_leave_success(client, manager_A, read_json):
-    response = client.get(
-        "user/my/leave/", headers={"Authorization": f"Bearer {manager_A}"}
-    )
-    expected = read_json(
-        "expected_responses/manager/leave/get_all_my_leave_manager_A.json"
-    )
+    response = client.get("user/my/leave/", headers={"Authorization": f"Bearer {manager_A}"})
+    expected = read_json("expected_responses/manager/leave/get_all_my_leave_manager_A.json")
     assert response.status_code == 200
     assert response.json() == expected
 
@@ -14,9 +10,7 @@ def test_manager_get_my_leave_by_year_success(client, manager_A, read_json):
     response = client.get(
         "user/my/leave/year/2025", headers={"Authorization": f"Bearer {manager_A}"}
     )
-    expected = read_json(
-        "expected_responses/manager/leave/get_my_leave_by_year_manager_A.json"
-    )
+    expected = read_json("expected_responses/manager/leave/get_my_leave_by_year_manager_A.json")
     assert response.status_code == 200
     assert response.json() == expected
 
@@ -26,15 +20,11 @@ def test_manager_get_my_leave_by_year_not_found(client, manager_A, read_json):
         "user/my/leave/year/2030", headers={"Authorization": f"Bearer {manager_A}"}
     )
     assert response.status_code == 404
-    assert response.json() == {
-        "detail": "Leave record not found for employee 2 in year 2030"
-    }
+    assert response.json() == {"detail": "Leave record not found for employee 2 in year 2030"}
 
 
-# -------------------------------------------------Test Admin API ---------------------------------------------------
-def test_manager_admin_access_get_employee_leave_by_empid_and_years_forbidden(
-    client, manager_A
-):
+# ------------------------------Test Admin API -------------------------------
+def test_manager_admin_access_get_employee_leave_by_empid_and_years_forbidden(client, manager_A):
     response = client.get(
         "admin/leaves/employee/1/year/2025",
         headers={"Authorization": f"Bearer {manager_A}"},
@@ -43,9 +33,7 @@ def test_manager_admin_access_get_employee_leave_by_empid_and_years_forbidden(
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_delete_employee_leave_by_empid_and_years_forbidden(
-    client, manager_A
-):
+def test_manager_admin_access_delete_employee_leave_by_empid_and_years_forbidden(client, manager_A):
     response = client.delete(
         "admin/leaves/employee/1/year/2025",
         headers={"Authorization": f"Bearer {manager_A}"},
@@ -54,9 +42,7 @@ def test_manager_admin_access_delete_employee_leave_by_empid_and_years_forbidden
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_get_employee_leaves_by_empid_forbidden(
-    client, manager_A, read_json
-):
+def test_manager_admin_access_get_employee_leaves_by_empid_forbidden(client, manager_A, read_json):
     response = client.get(
         "admin/leaves/employee/1", headers={"Authorization": f"Bearer {manager_A}"}
     )
@@ -83,12 +69,8 @@ def test_manager_admin_access_create_leave_forbidden(client, manager_A, read_jso
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_delete_leave_by_id_forbidden(
-    client, manager_A, read_json
-):
-    response = client.delete(
-        "admin/leaves/1", headers={"Authorization": f"Bearer {manager_A}"}
-    )
+def test_manager_admin_access_delete_leave_by_id_forbidden(client, manager_A, read_json):
+    response = client.delete("admin/leaves/1", headers={"Authorization": f"Bearer {manager_A}"})
 
     assert response.status_code == 403
     assert response.json() == {"detail": "Admin privileges required"}
