+++ tests/test_cases/test_manager_api/test_manager_payslips_api.py
@@ -1,18 +1,12 @@
-# -------------------------------------------------Test User API ---------------------------------------------------
+# -----------------------------------Test User API -----------------------------------
 def test_manager_get_all_my_payslips_success(client, manager_A, read_json):
-    response = client.get(
-        "/user/my/payslips/", headers={"Authorization": f"Bearer {manager_A}"}
-    )
-    expected = read_json(
-        "expected_responses/manager/payslips/get_all_my_payslips_manager_A.json"
-    )
+    response = client.get("/user/my/payslips/", headers={"Authorization": f"Bearer {manager_A}"})
+    expected = read_json("expected_responses/manager/payslips/get_all_my_payslips_manager_A.json")
     assert response.status_code == 200
     assert response.json() == expected
 
 
-def test_manager_get_my_payslips_by_year_and_month_success(
-    client, manager_A, read_json
-):
+def test_manager_get_my_payslips_by_year_and_month_success(client, manager_A, read_json):
     response = client.get(
         "/user/my/payslips/month/2025/1",
         headers={"Authorization": f"Bearer {manager_A}"},
@@ -24,9 +18,7 @@ def test_manager_get_my_payslips_by_year_and_month_success(
     assert response.json() == expected
 
 
-def test_manager_get_my_payslips_by_year_and_month_fot_found(
-    client, manager_A, read_json
-):
+def test_manager_get_my_payslips_by_year_and_month_fot_found(client, manager_A, read_json):
     response = client.get(
         "/user/my/payslips/month/2025/11",
         headers={"Authorization": f"Bearer {manager_A}"},
@@ -35,7 +27,7 @@ def test_manager_get_my_payslips_by_year_and_month_fot_found(
     assert response.json() == {"detail": "Payslip not found for this month"}
 
 
-# -------------------------------------------------Test Admin API ---------------------------------------------------
+# ------------------------------Test Admin API -------------------------------
 def test_manager_admin_access_create_payslips_forbidden(client, manager_A, read_json):
     payload = {
         "basic_amount": 1,
@@ -54,9 +46,7 @@ def test_manager_admin_access_create_payslips_forbidden(client, manager_A, read_
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_get_payslips_by_id_forbidden(
-    client, manager_A, read_json
-):
+def test_manager_admin_access_get_payslips_by_id_forbidden(client, manager_A, read_json):
     response = client.get(
         "/admin/payslips/employee/1", headers={"Authorization": f"Bearer {manager_A}"}
     )
@@ -64,9 +54,7 @@ def test_manager_admin_access_get_payslips_by_id_forbidden(
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_get_payslips_by_year_and_month_forbidden(
-    client, manager_A
-):
+def test_manager_admin_access_get_payslips_by_year_and_month_forbidden(client, manager_A):
     response = client.get(
         "/admin/payslips/month/2025/4", headers={"Authorization": f"Bearer {manager_A}"}
     )
@@ -74,9 +62,7 @@ def test_manager_admin_access_get_payslips_by_year_and_month_forbidden(
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_get_payslips_by_empid_year_and_month_forbidden(
-    client, manager_A
-):
+def test_manager_admin_access_get_payslips_by_empid_year_and_month_forbidden(client, manager_A):
     response = client.get(
         "/admin/payslips/employee/1/month/2025/4",
         headers={"Authorization": f"Bearer {manager_A}"},
@@ -85,9 +71,7 @@ def test_manager_admin_access_get_payslips_by_empid_year_and_month_forbidden(
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_delete_payslips_by_empid_year_and_month_forbidden(
-    client, manager_A
-):
+def test_manager_admin_access_delete_payslips_by_empid_year_and_month_forbidden(client, manager_A):
     response = client.delete(
         "/admin/payslips/employee/1/month/2025/4",
         headers={"Authorization": f"Bearer {manager_A}"},
@@ -96,11 +80,7 @@ def test_manager_admin_access_delete_payslips_by_empid_year_and_month_forbidden(
     assert response.json() == {"detail": "Admin privileges required"}
 
 
-def test_manager_admin_access_delete_payslips_by_id_forbidden(
-    client, manager_A, read_json
-):
-    response = client.delete(
-        "/admin/payslips/1", headers={"Authorization": f"Bearer {manager_A}"}
-    )
+def test_manager_admin_access_delete_payslips_by_id_forbidden(client, manager_A, read_json):
+    response = client.delete("/admin/payslips/1", headers={"Authorization": f"Bearer {manager_A}"})
     assert response.status_code == 403
     assert response.json() == {"detail": "Admin privileges required"}
